requirements.txt
# Automatically generated by https://github.com/damnever/pigar.

# C:\Projects\light-pipe\light_pipe\pipe\transformers.py: 9
# C:\Projects\light-pipe\light_pipe\utils\raster_trans.py: 1
GDAL == 3.5.1

# C:\Projects\light-pipe\light_pipe\gcs\gcs_io.py: 5
google_resumable_media == 2.3.3

# C:\Projects\light-pipe\light_pipe\pipe\transformers.py: 8
numpy == 1.23.0

# C:\Projects\light-pipe\light_pipe\pipe\storage.py: 5
# C:\Projects\light-pipe\light_pipe\scripts\gcs\download_items.py: 4
pandas == 1.4.3

# C:\Projects\light-pipe\light_pipe\tests\test_pipe.py: 1
# C:\Projects\light-pipe\light_pipe\tests\test_utils.py: 1
pytest == 7.1.2

# C:\Projects\light-pipe\light_pipe\scripts\gcs\download_items.py: 5,6
# C:\Projects\light-pipe\light_pipe\scripts\gcs\make_itemlist.py: 2
requests == 2.22.0
